Add check if uninstall was called from WordPress

+++ uninstall.php
@@ -6,6 +6,13 @@
  * @package         algolia-woo-indexer
  */
 
+/*
+*  If uninstall was not called from WordPress, then exit
+*/
+if ( ! defined( 'WP_UNINSTALL_PLUGIN' ) ) {
+	exit;
+}
+
 define( 'ALGOWOO_DB_OPTION', '_algolia_woo_indexer' );
 
 delete_option( ALGOWOO_DB_OPTION . '_application_id' );
